requirements.txt
flask
flask-cors
google-generative-ai
gunicorn